daemon: add handler for events
