+++ runtime/bpftime-daemon/CMakeLists.txt
@@ -6,7 +6,7 @@ add_bpf_skel_generating_target(bpftime_daemon_ebpf_skel ${CMAKE_CURRENT_BINARY_D
 
 add_dependencies(bpftime_daemon_ebpf_skel bpftime_daemon_ebpf_target)
 
-add_executable(bpftime_daemon bpf-mocker.cpp)
+add_executable(bpftime_daemon bpf-mocker.cpp handle_bpf_event.cpp)
 add_dependencies(bpftime_daemon bpftime_daemon_ebpf_skel libbpf)
 
 target_include_directories(bpftime_daemon PRIVATE ${CMAKE_CURRENT_BINARY_DIR} ${LIBBPF_INCLUDE_DIRS} ${CMAKE_CURRENT_SOURCE_DIR})
