+++ runtime/bpftime-daemon/bpf-mocker.cpp
@@ -17,6 +17,8 @@
 #include <linux/perf_event.h>
 #include "bpf-mocker-event.h"
 #include "bpf-mocker.skel.h"
+#include "daemon_config.hpp"
+#include "handle_bpf_event.hpp"
 
 /* Tune the buffer size and wakeup rate. These settings cope with roughly
  * 50k opens/sec.
@@ -29,15 +31,11 @@
 
 #define NSEC_PER_SEC 1000000000ULL
 
+using namespace bpftime;
+
 static volatile sig_atomic_t exiting = 0;
 
-static struct env {
-	pid_t pid;
-	uid_t uid;
-	bool verbose;
-	bool failed;
-	bool show_open;
-} env = { .uid = INVALID_UID };
+static struct env env = { .uid = INVALID_UID };
 
 const char *argp_program_version = "bpftime-daemon 0.1";
 const char *argp_program_bug_address = "https://github.com/eunomia-bpf/bpftime";
@@ -112,224 +110,12 @@ static void sig_int(int signo)
 	exiting = 1;
 }
 
-static int handle_open_events(const struct event *e)
-{
-	struct tm *tm;
-	char ts[32];
-	time_t t;
-	int fd, err;
-	if (!env.show_open) {
-		return 0;
-	}
-
-	/* prepare fields */
-	time(&t);
-	tm = localtime(&t);
-	strftime(ts, sizeof(ts), "%H:%M:%S", tm);
-	if (e->open_data.ret >= 0) {
-		fd = e->open_data.ret;
-		err = 0;
-	} else {
-		fd = -1;
-		err = -e->open_data.ret;
-	}
-
-	/* print output */
-	printf("OPEN %-6d %-16s %3d %3d ", e->pid, e->comm, fd, err);
-	printf("%s\n", e->open_data.fname);
-	return 0;
-}
-
-const char *bpf_cmd_strings[] = {
-	"BPF_MAP_CREATE",	 "BPF_MAP_LOOKUP_ELEM",
-	"BPF_MAP_UPDATE_ELEM",	 "BPF_MAP_DELETE_ELEM",
-	"BPF_MAP_GET_NEXT_KEY",	 "BPF_PROG_LOAD",
-	"BPF_OBJ_PIN",		 "BPF_OBJ_GET",
-	"BPF_PROG_ATTACH",	 "BPF_PROG_DETACH",
-	"BPF_PROG_TEST_RUN",	 "BPF_PROG_GET_NEXT_ID",
-	"BPF_MAP_GET_NEXT_ID",	 "BPF_PROG_GET_FD_BY_ID",
-	"BPF_MAP_GET_FD_BY_ID",	 "BPF_OBJ_GET_INFO_BY_FD",
-	"BPF_PROG_QUERY",	 "BPF_RAW_TRACEPOINT_OPEN",
-	"BPF_BTF_LOAD",		 "BPF_BTF_GET_FD_BY_ID",
-	"BPF_TASK_FD_QUERY",	 "BPF_MAP_LOOKUP_AND_DELETE_ELEM",
-	"BPF_MAP_FREEZE",	 "BPF_BTF_GET_NEXT_ID",
-	"BPF_MAP_LOOKUP_BATCH",	 "BPF_MAP_LOOKUP_AND_DELETE_BATCH",
-	"BPF_MAP_UPDATE_BATCH",	 "BPF_MAP_DELETE_BATCH",
-	"BPF_LINK_CREATE",	 "BPF_LINK_UPDATE",
-	"BPF_LINK_GET_FD_BY_ID", "BPF_LINK_GET_NEXT_ID",
-	"BPF_ENABLE_STATS",	 "BPF_ITER_CREATE",
-	"BPF_LINK_DETACH",	 "BPF_PROG_BIND_MAP",
-};
-
-const char* bpf_prog_type_strings[] = {
-    "BPF_PROG_TYPE_UNSPEC",
-    "BPF_PROG_TYPE_SOCKET_FILTER",
-    "BPF_PROG_TYPE_KPROBE",
-    "BPF_PROG_TYPE_SCHED_CLS",
-    "BPF_PROG_TYPE_SCHED_ACT",
-    "BPF_PROG_TYPE_TRACEPOINT",
-    "BPF_PROG_TYPE_XDP",
-    "BPF_PROG_TYPE_PERF_EVENT",
-    "BPF_PROG_TYPE_CGROUP_SKB",
-    "BPF_PROG_TYPE_CGROUP_SOCK",
-    "BPF_PROG_TYPE_LWT_IN",
-    "BPF_PROG_TYPE_LWT_OUT",
-    "BPF_PROG_TYPE_LWT_XMIT",
-    "BPF_PROG_TYPE_SOCK_OPS",
-    "BPF_PROG_TYPE_SK_SKB",
-    "BPF_PROG_TYPE_CGROUP_DEVICE",
-    "BPF_PROG_TYPE_SK_MSG",
-    "BPF_PROG_TYPE_RAW_TRACEPOINT",
-    "BPF_PROG_TYPE_CGROUP_SOCK_ADDR",
-    "BPF_PROG_TYPE_LWT_SEG6LOCAL",
-    "BPF_PROG_TYPE_LIRC_MODE2",
-    "BPF_PROG_TYPE_SK_REUSEPORT",
-    "BPF_PROG_TYPE_FLOW_DISSECTOR",
-    "BPF_PROG_TYPE_CGROUP_SYSCTL",
-    "BPF_PROG_TYPE_RAW_TRACEPOINT_WRITABLE",
-    "BPF_PROG_TYPE_CGROUP_SOCKOPT",
-    "BPF_PROG_TYPE_TRACING",
-    "BPF_PROG_TYPE_STRUCT_OPS",
-    "BPF_PROG_TYPE_EXT",
-    "BPF_PROG_TYPE_LSM",
-    "BPF_PROG_TYPE_SK_LOOKUP",
-    "BPF_PROG_TYPE_SYSCALL",
-};
-
-static int handle_bpf_event(const struct event *e)
-{
-	struct tm *tm;
-	char ts[32];
-	time_t t;
-
-	/* prepare fields */
-	time(&t);
-	tm = localtime(&t);
-	strftime(ts, sizeof(ts), "%H:%M:%S", tm);
-
-	const char *cmd_str = e->bpf_data.bpf_cmd >= (sizeof(bpf_cmd_strings) /
-						  sizeof(bpf_cmd_strings[0])) ?
-				  "UNKNOWN COMMAND" :
-				  bpf_cmd_strings[e->bpf_data.bpf_cmd];
-
-	/* print output */
-	printf("BPF %-6d %-16s %-16s %d\n", e->pid, e->comm, cmd_str, e->bpf_data.ret);
-	return 0;
-}
-
-#define PERF_TYPE_MAX_ID 16
-const char* perf_type_id_strings[PERF_TYPE_MAX_ID] = {
-    [PERF_TYPE_HARDWARE]     = "PERF_TYPE_HARDWARE",
-    [PERF_TYPE_SOFTWARE]     = "PERF_TYPE_SOFTWARE",
-    [PERF_TYPE_TRACEPOINT]   = "PERF_TYPE_TRACEPOINT",
-    [PERF_TYPE_HW_CACHE]     = "PERF_TYPE_HW_CACHE",
-    [PERF_TYPE_RAW]          = "PERF_TYPE_RAW",
-    [PERF_TYPE_BREAKPOINT]   = "PERF_TYPE_BREAKPOINT",
-};
-
-static int handle_perf_event(const struct event *e)
-{
-	struct tm *tm;
-	char ts[32];
-	time_t t;
-
-	/* prepare fields */
-	time(&t);
-	tm = localtime(&t);
-	strftime(ts, sizeof(ts), "%H:%M:%S", tm);
-
-	const char *type_id_str = e->perf_event_data.attr.type >= (sizeof(perf_type_id_strings) /
-						  sizeof(perf_type_id_strings[0])) ?
-				  "UNKNOWN TYPE" :
-				  perf_type_id_strings[e->perf_event_data.attr.type];
-
-	/* print output */
-	printf("PERF %-6d %-16s %-8s %d\n", e->pid, e->comm, type_id_str, e->perf_event_data.pid);
-	return 0;
-}
-
-
-/*
- * this function is expected to parse integer in the range of [0, 2^31-1] from
- * given file using scanf format string fmt. If actual parsed value is
- * negative, the result might be indistinguishable from error
- */
-static int parse_uint_from_file(const char *file, const char *fmt)
-{
-	int err, ret;
-	FILE *f;
-
-	f = fopen(file, "re");
-	if (!f) {
-		err = -errno;
-		fprintf(stderr, "failed to open '%s\n", file);
-		return err;
-	}
-	err = fscanf(f, fmt, &ret);
-	if (err != 1) {
-		err = err == EOF ? -EIO : -errno;
-		fprintf(stderr,"failed to parse '%s'\n", file);
-		fclose(f);
-		return err;
-	}
-	fclose(f);
-	return ret;
-}
-
-static int determine_kprobe_perf_type(void)
-{
-	const char *file = "/sys/bus/event_source/devices/kprobe/type";
-
-	return parse_uint_from_file(file, "%d\n");
-}
-
-static int determine_uprobe_perf_type(void)
-{
-	const char *file = "/sys/bus/event_source/devices/uprobe/type";
-
-	return parse_uint_from_file(file, "%d\n");
-}
-
-
-static int handle_load_bpf_prog_event(const struct event *e)
-{
-	struct tm *tm;
-	char ts[32];
-	time_t t;
-
-	/* prepare fields */
-	time(&t);
-	tm = localtime(&t);
-	strftime(ts, sizeof(ts), "%H:%M:%S", tm);
-
-
-	const char *prog_type_str = e->bpf_loaded_prog.type >= (sizeof(bpf_prog_type_strings) /
-						  sizeof(bpf_prog_type_strings[0])) ?
-				  "UNKNOWN PROG TYPE" :
-				  bpf_prog_type_strings[e->bpf_loaded_prog.type];
-	
-	/* print output */
-	printf("LOAD %-6d %-16s %-16s %-16s %d\n", e->pid, e->comm, prog_type_str, e->bpf_loaded_prog.prog_name, e->bpf_loaded_prog.insn_cnt);
-	return 0;
-}
 
 static int handle_event_rb(void *ctx, void *data, size_t data_sz)
 {
 	const struct event *e = (const struct event *)data;
-	switch (e->type) {
-	case SYS_OPEN:
-		return handle_open_events(e);
-		break;
-	case SYS_PERF_EVENT_OPEN:
-		return handle_perf_event(e);
-		break;
-	case SYS_BPF:
-		return handle_bpf_event(e);
-		break;
-	case BPF_PROG_LOAD_EVENT:
-		return handle_load_bpf_prog_event(e);
-		break;
-	}
+	bpf_event_handler handler(env);
+
 	return 0;
 }
 
@@ -365,9 +151,6 @@ int main(int argc, char **argv)
 	/* initialize global data (filtering options) */
 	obj->rodata->target_pid = env.pid;
 	obj->rodata->disable_modify = true;
-
-	perf_type_id_strings[determine_uprobe_perf_type()] = "PERF_TYPE_UPROBE";
-	perf_type_id_strings[determine_kprobe_perf_type()] = "PERF_TYPE_KPROBE";
 
 	err = bpf_mocker_bpf__load(obj);
 	if (err) {
