--- runtime/bpftime-daemon/handle_event.hpp
@@ -1,5 +1,0 @@
-#ifndef BPFTIME_HANDLE_EVENT_HPP
-#define BPFTIME_HANDLE_EVENT_HPP
-
-
-#endif // BPFTIME_HANDLE_EVENT_HPP
